Bugfix - Não estava deletando as informações do cover no banco de dados, acabei esquecendo de colocar o método que exclui o mesmo

+++ app/Http/Controllers/Admin/Blog/PostController.php
@@ -119,8 +119,10 @@ class PostController extends Controller
 
     $this->postRepository->delete($post);
 
-    if (isset($post->cover->url) || Storage::disk('public')->exists($post->cover->url)) {
+    if (isset($post->cover->url) && Storage::disk('public')->exists($post->cover->url)) {
       Storage::disk('public')->delete($post->cover->url);
+
+      $post->cover()->delete();
     }
 
     return response()->json([], Response::HTTP_NO_CONTENT);
